Recipe editing/deleting permissions, and formatting

+++ essen/recipes/views.py
@@ -9,9 +9,14 @@ from django.urls import reverse
 
 # Create your views here.
 
-class DetailView(generic.DetailView):
-    model = Recipe
+def detail(request, pk):
     template_name = 'recipes/detail.html'
+    steward = False
+    if request.user.is_authenticated and check_if_steward(request.user):
+        steward = True
+
+    return render(request, template_name, {"recipe": get_object_or_404(Recipe, pk=pk),
+                                                  "steward": steward})
 
 def view_recipes(request):
     d = dict(request.GET.iterlists())
@@ -48,31 +53,45 @@ def add_recipe(request):
 
 
 def submit_recipe(request):
-    d = dict(request.POST.iterlists())
-    r = Recipe(recipe_name=d['recipe_name'][0], directions=d['directions'][0], serving_size=int(d['serving_size'][0]))
-    r.save()
+    if request.user.is_authenticated:
+        d = dict(request.POST.iterlists())
+        r = Recipe(recipe_name=d['recipe_name'][0], directions=d['directions'][0], serving_size=int(d['serving_size'][0]))
+        r.save()
 
-    for i in range(len(d['Ingredient'])):
-        Ingredient(recipe=r, ingredient_name=d['Ingredient'][i], units=d['Unit'][i],
-                       quantity=float(d['Quantity'][i])).save()
+        if 'ingredient' in d:
+            for i in range(len(d['Ingredient'])):
+                Ingredient(recipe=r, ingredient_name=d['Ingredient'][i], units=d['Unit'][i],
+                               quantity=float(d['Quantity'][i])).save()
 
     return HttpResponseRedirect(reverse('recipes:detail', args=[r.id]))
 
 
 def submit_edit(request, recipe_id):
-    recipe = get_object_or_404(Recipe, pk=recipe_id)
-    for ingredient in recipe.ingredient_set.all():
-        ingredient.delete()
-    d = dict(request.POST.iterlists())
+    if request.user.is_authenticated and check_if_steward(request.user):
+        recipe = get_object_or_404(Recipe, pk=recipe_id)
+        for ingredient in recipe.ingredient_set.all():
+            ingredient.delete()
+        d = dict(request.POST.iterlists())
 
-    recipe.serving_size = d['serving_size'][0]
-    recipe.recipe_name = d['recipe_name'][0]
-    recipe.directions = d['directions'][0]
-    recipe.save()
+        recipe.serving_size = d['serving_size'][0]
+        recipe.recipe_name = d['recipe_name'][0]
+        recipe.directions = d['directions'][0]
+        recipe.save()
 
-    for i in range(len(d['ingredient'])):
-        Ingredient(recipe=recipe, ingredient_name=d['ingredient'][i], units=d['units'][i],
-                       quantity=float(d['quantity'][i])).save()
+        if 'ingredient' in d:
+            for i in range(len(d['ingredient'])):
+                Ingredient(recipe=recipe, ingredient_name=d['ingredient'][i], units=d['units'][i],
+                               quantity=float(d['quantity'][i])).save()
 
     return HttpResponseRedirect(reverse('recipes:detail', args=[recipe_id]))
 
+def delete(request, recipe_id):
+    if request.user.is_authenticated and check_if_steward(request.user):
+        recipe = get_object_or_404(Recipe, pk=recipe_id)
+        recipe.delete()
+
+    return HttpResponseRedirect(reverse('recipes:index'))
+
+
+def check_if_steward(user):
+    return user.groups.all().filter(name="stewards").count() > 0
